UPSTREAM: phy: Add Embedded DisplayPort and DisplayPort submodes
In some cases, a DP PHY needs to be configured to work in eDP mode.
So add submodes for both DP and eDP so they can be used by the
controllers for specifying the mode the PHY should be configured in.

Change-Id: I130b491edcf2624697489b18a3ab24add535840f
Signed-off-by: Abel Vesa <abel.vesa@linaro.org>
Reviewed-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Link: https://lore.kernel.org/r/20240324-x1e80100-phy-edp-compatible-refactor-v5-1-a0db5f3150bc@linaro.org
Signed-off-by: Vinod Koul <vkoul@kernel.org>
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>
(cherry picked from commit 368d67dab4cc4a3ffd39fbd062b2f5796cdbb37b)