drm/bridge: synopsys: dw-hdmi-qp: Keep display state if userspace doesn't disable display pipeline after hdmi plug out.
However, the hdmi 2.0 spec requires that some flags are
set in the sink scdc regs, and those flags are very likely
to be reset when the cable has been disconnected. This will
thus result in a blank display, even if the display pipeline
configuration hasn't been modified or is in the exact
same state.

The solution we've had so far is to enable the scrambling
and high ratio related bits again on reconnection, but the
hdmi 2.0 spec (Section 6.1.3.1 - Scrambling Control) requires
that the scdc bit is set before sending any scrambled video
signal. Using that solution thus breaks that expectation.

So we need disable hdmi signal output when hdmi plug out.
writing scdc scrambling and high ratio regs when hdmi
plug in then enable hdmi signal output finally.

Change-Id: I422455e1d5b60c3fd4a3ef35ccb46de9fb0f28f8
Signed-off-by: Algea Cao <algea.cao@rock-chips.com>