drm/rockchip: logo: call drm_atomic_bridge_chain_check() bridge in mode fixup
For some connector drivers, such as dw-dp, &drm_bridge_funcs.atomic_check()
helps to select output_mode, color_range and so on, which will be
determined in the following &drm_encoder_helper_funcs.atomic_check().

In order to get the exact display mode for the mode comparison
between U-boot and Kernel, add the full bridge atomic check process
in rockchip_drm_mode_fixup().

Without this patch, the mode comparison will fail if the DP logo
display is enabled in YUV420 mode for RK3576, because the display
mode of uboot has been fixed while the display mode of kernel has
not, due to the incorrect conn_state->output_mode.

The related workaround is:

/*
 * For RK3576 YUV420 output, hden signal introduce one cycle delay,
 * so we need to adjust hfp and hbp to compatible with this design.
 */
if (vop2->version == VOP_VERSION_RK3576 &&
    conn_state->output_mode == ROCKCHIP_OUT_MODE_YUV420) {
	mode->crtc_hsync_start += 2;
	mode->crtc_hsync_end += 2;
}

Fixes: c4642391b1fe ("drm/rockchip: vop2: adjust hfp and hbp for YUV420 output")
Change-Id: I3193bfa2bca7c14f5c7308f94e7d23160618be0c
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>