drm/rockchip: drv: support get old crtc and new crtc by encoder
For all the display interface on the Soc, for example, DP, HDMI,
DSI, DPI. All of these display interface will register the drm encoder
which will encode the match the display interface requirement. In some
case, we may need check the encoder's crtc change or not to do some
work.
When userspace do a commit work, a connect may attach to a new crtc,
or unattach to a old crtc, and choose a best encoder. The connector
can know whether the new crtc and the old crtc it attach is the same
or know. But its best encoder can't know.
Here, we get the connector whose old_state.best_encoder is the target
encoder. Then the old_state.crtc of this connector is the encoder's old
crtc. We get the connector whose new_state.best_encoder is the target
encoder. Then the new_state.crtc if this connector is the encoder's new
crtc.

Change-Id: Id9e5f27b9dcf368473b48f76e3cad375276eadfb
Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>