drm/bridge: analogix_dp: add support forced switching to DP mode
Some bridges support to convert DP into display interfaces that support
the panel, such as RGB, LVDS, MIPI and so on. For instance, the LT7211B
and LT7911D from Lontium can offer this functionality.

Therefore, we add a property of the eDP node to support forced switching
to DP mode, and the DT setting may be like:

&edp0 {
	dp-mode;
};

Change-Id: I0f7c2598f86fb9d5ebedc945e3acd66388215d16
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>