drm/rockchip: vop2: remove the plane number check in vop2_plane_mask_check()
The &vop2_data.nr_layers represents the max number of planes that can
be assigned to one VP. Therefore, when checking the max plane number for
VOP2, it may be incorrect to use the &vop2_data.nr_layers.

And the comparison of assigned plane_mask with &vop2_data.plane_mask_base
can already confirm whether every plane has been assigned to one of all
the VPs.

Fixes: 00424a589d3d ("drm/rockchip: vop2: add support to assign plane mask by default")
Change-Id: I9c2056dea42897150ef155c7e6fd0ad3a43c8132
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>