drm/rockchip: vop2: fix the default plane_make configurations for RK3588
The relationships between the main window and the splice window are:

Main     | Splice
Cluster0 | Cluster1
Cluster2 | Cluster3
Esmart0  | Esmart1
Esmart2  | Esmart3

The VP0 and VP1 should be used in combination when the display mode is
over 4k. The main window should attach to the VP0, while the splice
window should attach to the other.

Without this patch, the default plane_mask may not meet the above
requirement when the plane_mask is not assigned in DTS.

Change-Id: Ia676f519ce26d579b1066841715fe8678dc15852
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>