ASoC: rockchip: sai: Fix the slot mask bitmap
The tx and rx masks use a 1 bit for an active slot as default,
and the default active bits are at the LSB of the masks.

The explicit masks are given as array of integers, where the first
number presents bit-0 (LSB), second presents bit-1, etc. Any non zero
number is considered 1 and 0 is 0.

So, we need do a XOR to get the masked bitmap.

Fixes: e0d4c336d281 ("ASoC: rockchip: sai: Add support for slot mask cfg")

Signed-off-by: Sugar Zhang <sugar.zhang@rock-chips.com>
Change-Id: I7f9baea58a4f6bea3e99f97ef9161ac94152a6f5