Mali: valhall: from ARM: Fix UAF issue of user IO pages after group fatal error
The user IO page mappings were not unmapped properly when a group is
terminated, which could lead to the pages being accessed after
termination. With this change, the driver shall remap a group's shared
page to the dummy page during termination to ensure that accesses to
it will not cause the application to crash, but also these accesses will
be rendered harmless.

For RK, this modification is derived from the kernel_patch_from_Eric_250412
provided by ARM in support case 03504455.
It is originally intended to address the issue
that the GKI of kernel 5.10 does not provide 'zap_vma_ptes'
which the r54 driver depends on.

Change-Id: I28a6946436343d2b44fe7e4df9b4bb8bbd3bafe1
Signed-off-by: Zhen Chen <chenzhen@rock-chips.com>