rockchp/drm: vop2: limit vrr work condition for active and mode change
In the active change and mode change case, the refresh_rate_change
should be set true to reconfig the reuqesut rate after active
change or mode change.
In some case, the active change or mode change is triggered, but
a valid request refresh rate is not config, which will also cause the
refresh_rate_change be set true. This is a wrong operation.
So in the active change or mode change case, It's also need to
check the whether request refresh rate is valid or not.

Fix error log at system bootup:
[drm:vop2_crtc_atomic_begin] *ERROR* invalid rate:0

Change-Id: I93f6f36a64302ddb0f38264377bb4659d8aaa5a2
Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>