drm/rockchip: drv: Only set clk loader protect when route node exist
rockchip_clocks_loader_protect() always call in arch_initcall_sync().
However, rockchip_clocks_loader_unprotect() is not called when the
DRM driver is not loaded, which assumes that these clocks will always
be enabled.

This may cause the clock to be turned on accidentally. In the case of
ebc, for example, when the ebc driver is enabled and the DRM driver
is not loaded, rockchip_clocks_loader_protect() will enable ebc clk.
In addition, rockchip_clocks_loader_unprotect() will not be called,
meaning that the ebc clock will always be on. However, the ebc driver
will also control the ebc clk, but it will not be able to actually
turn off the clk. The dmc driver will check the status of the clk in
the CRU while working, and it will always find that the clk is on.
This will cause a wrong judgment on the working status of the ebc.

This patch adds constraints to rockchip_clocks_loader_protect() to
protect clocks only when route nodes are available, which means that
the logo is already enabled during the u-boot stage, so it makes
sense to perform clock loader protect.

Change-Id: Ib9efbade547181e86516d19637995569a751fade
Signed-off-by: Chaoyi Chen <chaoyi.chen@rock-chips.com>