ubi: Change to fill ec_hdr and vid_hdr redundant space with ff data on rockchip
These hdr data is place in first two pages in the front of flash block,
most of the data is filled with 0 which may result in 0/1 data unbalance
and finally accelerate the data bit flip process. So using ff redundant
data pattern for data balancing is more stable for nand devices.

Change-Id: I77a9116a4575b94bc6e4da5334db4eaf3d8bcb83
Signed-off-by: Jon Lin <jon.lin@rock-chips.com>